Большие переделки Наметил контроллеры Переработал структуры, но все равно еще не нравятся они мне. Буду дальше эксперементировать. Фронтенд пока заброшу, займусь им позже

+++ Pathfinder.Core/Entities/Shop/ShopsProduct.cs
@@ -7,6 +7,6 @@ namespace Pathfinder.Core.Entities.Shop
     {
         public virtual Shop Shop { get; set; }
         public virtual Item Item { get; set; }
-        public int Count { get; set; }
+        public short Count { get; set; }
     }
 }
